+++ tests/system/test_enrollment/test_enrollment.py
@@ -133,20 +133,20 @@ def modify_ip_address_conf(test_case):
     address_ip = ''
     message_ip_manager = ''
     message_ip_agent = ''
-    message_dns_manager = ''
+    message_address_manager = ''
     final_message = ''
 
     if test_case['ip_type'] == 'ipv4':
         address_ip = network['manager_network'][0]
-        message_ip_manager = address_ip
+        message_address_manager = message_ip_manager = address_ip
         message_ip_agent = network['agent_network'][0]
     elif test_case['ip_type'] == 'ipv6':
         address_ip = network['manager_network'][1]
-        message_ip_manager = format_ipv6_long(address_ip)
+        message_address_manager = message_ip_manager = format_ipv6_long(address_ip)
         message_ip_agent = format_ipv6_long(network['agent_network'][1])
     elif test_case['ip_type'] == 'dns':
         address_ip = 'wazuh-manager'
-        message_dns_manager = f"{address_ip}\\/"
+        message_address_manager = address_ip
         if test_case['ipv6_enabled'] == 'yes':
             if 'ipv4' in test_case['manager_network'] or 'ipv4' in test_case['agent_network']:
                 message_ip_manager = f"{network['manager_network'][0]}"
@@ -162,9 +162,9 @@ def modify_ip_address_conf(test_case):
                                                         f"<address>{address_ip}</address>")
     host_manager.modify_file_content(host='wazuh-agent1', path='/var/ossec/etc/ossec.conf',
                                      content=new_configuration)
-    message_dns_manager = message_dns_manager.replace(r'-', r'\\-')
-    message_with_manager_dns = messages.replace(r'MANAGER_DNS\\/', message_dns_manager)
-    message_with_manager_ip = message_with_manager_dns.replace('MANAGER_IP', message_ip_manager)
+    message_address_manager = message_address_manager.replace(r'-', r'\\-')
+    message_with_manager_address = messages.replace(r'MANAGER_ADDRESS', message_address_manager)
+    message_with_manager_ip = message_with_manager_address.replace('MANAGER_IP', message_ip_manager)
     final_message = message_with_manager_ip.replace('AGENT_IP', message_ip_agent)
     write_file(messages_path, final_message)
 
@@ -224,7 +224,7 @@ def test_agent_enrollment(test_case, get_ip_directions, configure_network, modif
 
     # Start the agent enrollment process by restarting the wazuh-agent
     host_manager.control_service(host='wazuh-manager', service='wazuh', state="restarted")
-    host_manager.get_host('wazuh-agent1').ansible('command', f'service wazuh-agent restart', check=False)
+    host_manager.control_service(host='wazuh-agent1', service='wazuh', state="restarted")
 
     # Run the callback checks for the ossec.log
     HostMonitor(inventory_path=inventory_path,