+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Release report: TBD
 
 ### Changed
 
+- Update enrollment logs in system test ([#4442](https://github.com/wazuh/wazuh-qa/pull/4442)) \- (Tests)
 - Fix one_manager_agent environment provisioning by packages for system tests ([#4438](https://github.com/wazuh/wazuh-qa/pull/4438)) \- (Framework)
 
 ## [4.5.1] - 24-08-2023